makes optimization arguments available for TestVC4C
