+++ src/tools/sources.list
@@ -2,4 +2,5 @@ target_sources(${VC4C_LIBRARY_NAME}
   PRIVATE
     ${CMAKE_CURRENT_LIST_DIR}/Emulator.cpp
     ${CMAKE_CURRENT_LIST_DIR}/Emulator.h
+    ${CMAKE_CURRENT_LIST_DIR}/options.cpp
 )
